+++ src/VFXToolbox.cs
@@ -17,12 +17,14 @@ public static class VFX
     public static GameObject LaserSightPrefab;
     public static GameObject MiniPickup;
     public static GameObject MasterySigil;
+    public static GameObject BasicReticle;
 
     public static void Init()
     {
         LaserSightPrefab = LoadHelper.LoadAssetFromAnywhere("assets/resourcesbundle/global vfx/vfx_lasersight.prefab") as GameObject;
         MiniPickup = VFX.Create("mini_pickup", 12, loops: false, anchor: Anchor.MiddleCenter);
         MasterySigil = VFX.Create("mastery_sigil", fps: 2);
+        BasicReticle = VFX.Create("basic_reticle", fps: 2);
     }
 
     /// <summary>Register an animation for a VFX object</summary>