+++ changelog.md
@@ -9,6 +9,8 @@
 	- Added animations for dropping current juggle combo
 	- Added projectile impact sounds
 	- Added circus music :D
+- Added targeting reticle to Iron Maid
+- Made Iron Maid projectiles launch towards target enemy's current position at time of launch (rather than their old position at time of reload)
 
 ## 1.14.0 (2024-06-10)